requirements.txt
numpy~=1.18.1
lmdb~=0.98
six~=1.14.0
click~=7.1.2
loguru~=0.5.1
tensorflow-gpu=2.2.0
easydict
anyconfig
pyyaml
pytest
pillow
cython
notifiers
